Add color options for Border Play and Mute
Personally i've always wanted a different color for the > play
indicator to make it easier to read, in my mind it made sense to
do the - mute indicator at the same time. In the default configs
these are still mapped to the same color as cursor and normal.

In adition someone on Discord was after the ability to change the
dialog border so I added that while I was there.

This commit also cleans up some whitespace in the parts of the
code and docs that it touched.