Improve Midi Input Part 1 (#142)
* MidiService: Use outDevice instead of device

This is so that we can add a

Previously the midi in was only started if the midi device was
present in the config. With this change it is now started and
updated whenever the project settings change. This includes
starting input when a project is loaded if a device is set in the
project.

* MidiInDevice: Cleanup codebase

* 4 spaces everywhere
* Remove extra semicolons
* No space before semicolon

* Fix Midi In Start Stop

Now that midi devices are created correctly I could actually test
this.

* RG35XXPLUS: Fix mapping file

rshoulder and lshoulder were the wrong way

* RG35XXPLUS: Clean up mapping.xml

* Use 4 spaces for indent everywhere
* Put everything in a more sensible order

* clang-format run